refactor: 도메인 중심 구조 패키지 수정 완료
- 모든 파일의 패키지 선언 및 import 경로 수정
- GlobalExceptionHandler를 common/exception으로 이동
- 불필요한 파일 제거
- 도메인 중심 구조로 완전히 재구성됨
