+++ src/main/java/com/melog/melog/emotion/adapter/emotion/EmotionKeywordJpaRepository.java
@@ -1,4 +1,4 @@
-package com.melog.melog.adapter.out.persistence.emotion;
+package com.melog.melog.emotion.adapter.emotion;
 
 import com.melog.melog.emotion.domain.EmotionKeyword;
 import com.melog.melog.emotion.domain.EmotionRecord;
