+++ src/main/java/com/melog/melog/emotion/adapter/emotion/EmotionScorePersistenceAdapter.java
@@ -1,6 +1,6 @@
-package com.melog.melog.adapter.out.persistence.emotion;
+package com.melog.melog.emotion.adapter.emotion;
 
-import com.melog.melog.application.port.out.EmotionScorePersistencePort;
+import com.melog.melog.emotion.application.EmotionScorePersistencePort;
 import com.melog.melog.emotion.domain.EmotionRecord;
 import com.melog.melog.emotion.domain.EmotionScore;
 import com.melog.melog.emotion.domain.EmotionType;
